+++ packages/ws-worker/src/channels/worker-queue.ts
@@ -1,25 +1,22 @@
+import EventEmitter from 'node:events';
 import { Socket as PhxSocket } from 'phoenix';
 import { WebSocket } from 'ws';
 
 import generateWorkerToken from '../util/worker-token';
-import type { Socket, Channel } from '../types';
 
-type SocketAndChannel = {
-  socket: Socket;
-  channel: Channel;
-};
+import type { Logger } from '@openfn/logger';
+import type { Channel } from '../types';
 
 const connectToWorkerQueue = (
   endpoint: string,
   serverId: string,
   secret: string,
+  logger: Logger,
   SocketConstructor = PhxSocket
 ) => {
-  return new Promise<SocketAndChannel>(async (done, reject) => {
-    // TODO does this token need to be fed back anyhow?
-    // I think it's just used to connect and then forgotten?
-    // If we reconnect we need a new token I guess?
-    const token = await generateWorkerToken(secret, serverId);
+  const events = new EventEmitter();
+
+  generateWorkerToken(secret, serverId).then((token) => {
     // @ts-ignore ts doesn't like the constructor here at all
     const socket = new SocketConstructor(endpoint, {
       params: { token },
@@ -28,36 +25,40 @@ const connectToWorkerQueue = (
 
     let didOpen = false;
 
-    // TODO need error & timeout handling (ie wrong endpoint or endpoint offline)
-    // Do we infinitely try to reconnect?
-    // Consider what happens when the connection drops
-    // Unit tests on all of these behaviours!
     socket.onOpen(() => {
       didOpen = true;
 
-      // join the queue channel
-      // TODO should this send the worker token?
       const channel = socket.channel('worker:queue') as Channel;
 
       channel
         .join()
         .receive('ok', () => {
-          done({ socket, channel });
+          logger.debug('Connected to worker queue socket');
+          events.emit('connect', { socket, channel });
         })
         .receive('error', (e: any) => {
-          console.log('ERROR', e);
+          logger.error('ERROR', e);
         })
         .receive('timeout', (e: any) => {
-          console.log('TIMEOUT', e);
+          logger.error('TIMEOUT', e);
         });
     });
 
-    // if we fail to connect
+    // On close, the socket will try and reconnect itself
+    // Forever, so far as I can tell
+    socket.onClose((_e: any) => {
+      logger.debug('queue socket closed');
+      events.emit('disconnect');
+    });
+
+    // if we fail to connect, the socket will try to reconnect
+    /// forever (?) with backoff
     socket.onError((e: any) => {
       // If we failed to connect, reject the promise
       // The server will try and reconnect itself.s
       if (!didOpen) {
-        reject(e);
+        events.emit('error', e.message);
+        didOpen = false;
       }
       // Note that if we DID manage to connect once, the socket should re-negotiate
       // wihout us having to do anything
@@ -65,6 +66,8 @@ const connectToWorkerQueue = (
 
     socket.connect();
   });
+
+  return events;
 };
 
 export default connectToWorkerQueue;