+++ packages/ws-worker/src/types.d.ts
@@ -1,5 +1,5 @@
 import { SanitizePolicies } from '@openfn/logger';
-import type { Socket as PhxSocket, Channel as PhxChannel } from 'phoenix';
+import type { Channel as PhxChannel } from 'phoenix';
 
 export { Socket };